extract encode to braille method into its own module
