+++ lib/night_writer.rb
@@ -1,6 +1,9 @@
 require_relative 'alphabet'
+require_relative 'encoder'
 
 class NightWriter
+
+  include Encoder
 
   def initialize
     @alphabet = Alphabet.new
@@ -8,21 +11,6 @@ class NightWriter
 
   def lookup(character, position)
     @alphabet.braille_letter_hash[character].chars[position]
-  end
-
-  def encode_to_braille(plain)
-    output = []
-    [0,2,4].each do |offset|
-      plain.chars.each do |letter|
-        if letter == letter.upcase
-          output << lookup(:capitalize, offset) << lookup(:capitalize, offset + 1)
-          letter = letter.downcase
-        end
-        output << lookup(letter, offset) << lookup(letter, offset + 1)
-      end
-      output << "\n"
-    end
-    encoded = output.join
   end
 
   def encode_from_braille(braille)
@@ -56,4 +44,3 @@ class NightWriter
     output.join
   end
 end
-
